+++ src/features/cfgActionProvider.ts
@@ -18,9 +18,11 @@ export function activate(context: vscode.ExtensionContext) {
     if (match && (checked = utils.hasGraphicsFile(modPaths, match[2], annoRda)).length > 0) {
       const index = line.text.indexOf(match[2]);
       const range = new vscode.Range(line.lineNumber, index, line.lineNumber, index + match[2].length);
+
+      const allPaths = annoRda ? [annoRda, ...modPaths] : modPaths;
   
       const diagnostic = new vscode.Diagnostic(range,
-        `File seems to be missing. Checked paths:\n` + checked.join('\n'),
+        `File seems to be missing.\nChecked paths:\n${allPaths.join('\n')}\nChecked patterns:\n${checked.join('\n')}`,
         vscode.DiagnosticSeverity.Warning);
       return diagnostic;
     }
@@ -37,10 +39,11 @@ export function activate(context: vscode.ExtensionContext) {
     const config = vscode.workspace.getConfiguration('anno', doc.uri);
     const checkFileNames = config.get('checkFileNames');
     const annoRda: string | undefined = config.get('rdaFolder');
+    const modsFolder: string | undefined = config.get('modsFolder');
 
     const diagnostics: vscode.Diagnostic[] = [];
 
-    const modPaths = utils.searchModPaths(doc.uri.fsPath);
+    const modPaths = utils.searchModPaths(doc.uri.fsPath, modsFolder);
 
     for (let lineIndex = 0; lineIndex < doc.lineCount; lineIndex++) {
       const lineOfText = doc.lineAt(lineIndex);